Refactor habit form store import paths to use centralized store location
