+++ app/habits/create/time-of-day.tsx
@@ -3,7 +3,7 @@ import { Pressable, ScrollView, View } from "react-native";
 import { Text } from "~/components/ui/text";
 import { fontFamily } from "~/lib/font";
 import { Check } from "~/lib/icons/Check";
-import { useHabitFormStore } from "./habit-store";
+import { useHabitFormStore } from "~/stores/habit-store";
 import { cn } from "~/lib/utils";
 import { useShallow } from "zustand/react/shallow";
